Added Check/Create Dir button
